Add basic 3D map support and launcher
